Relationships

+++ database/migrations/2019_10_10_074559_create_replies_table.php
@@ -17,7 +17,7 @@ class CreateRepliesTable extends Migration
             $table->bigIncrements('id');
             $table->text('body');
 
-            $table->integer('question_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
             $table->integer('user_id');
 
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
